requirements.txt
alembic==1.7.7
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.0.12
click==8.0.4
Deprecated==1.2.13
Flask==2.0.3
Flask-Caching==1.10.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.5.1
Faker==1.16.0
greenlet==1.1.3
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
migrate==0.3.8
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-dotenv==0.20.0
redis==4.3.4
requests==2.27.1
SQLAlchemy==1.4.41
typing_extensions==4.1.1
urllib3==1.26.12
Werkzeug==2.0.3
wrapt==1.14.1
zipp==3.6.0
randomuser==1.6
